requirements.txt
atlassian-python-api==3.12.1
